feat!: modularize cambridge dict, refactor reset()
Started modularizing the background script to make it more sustainable.
Made the Cambridge Dictionary resource a separate module and changed
how resource reset works.

Now, if you want to reset all resources, you don't provide the
settings.reset() method with any arguments, and if you want to reset a
particular resource, then you need to pass the resource's ID to the
method.

At this moment, the underlying mechanism of the refactored method is
the same as before for all resources except Cambridge Dictionary. But I
intend to refactor all resources this way.